requirements.txt
streamlit
python-dotenv
truststore
requests
PyYAML